Index Handlers

The task of an index handler is to provide a mapping between the system state and the way it is stored in memory, usually as a multidimensional array. The standard approach is to represent the states of a system with $s$ reactions as an $s$-dimensional array and have the index $(i_1, \ldots, i_s)$ correspond to the state $(n_1 = i_1, \ldots, n_s = i_s)$. This is implemented by the class NaiveIndexHandler, which accepts an offset argument to deal with Julia's 1-based indexing (so the Julia index $(1,\ldots,1)$ corresponds to the state with no molecules). For systems with conservation laws the DefaultIndexHandler class generally stores the data more efficiently.

See the internal API on how to define your own IndexHandler type.

FiniteStateProjection.NaiveIndexHandler — Type

struct NaiveIndexHandler <: AbstractIndexHandler
    offset::Int
end

Basic index handler that stores the state of a system with s species in an s-dimensional array. The offset parameter denotes the offset by which the array is indexed (defaults to 1 in Julia). Use OffsetArrays.jl to enable 0-based indexing.

This is the simplest index handler, but it will not be optimal if some states cannot be reached from the initial state, e.g. due to the presence of conservation laws. It is generally better to use DefaultIndexHandler, which will automatically elide species where possible.

Constructors: NaiveIndexHandler([sys::FSPSystem, offset::Int=1])

See also: DefaultIndexHandler

FiniteStateProjection.DefaultIndexHandler — Type

struct DefaultIndexHandler <: AbstractIndexHandler

More efficient index handler that improves upon NaiveIndexHandler by eliminating variables whose abundances can be computed from other variables using conservation laws. Describes the system using a subset of the original species which can be obtained via reducedspecies. Reduces the dimensionality of the FSP by the number of conservation laws in the system.

Constructors: DefaultIndexHandler(sys::FSPSystem[, offset::Int=1])

See also: reducedspecies, elidedspecies, NaiveIndexHandler

FiniteStateProjection.reducedspecies — Function

reducedspecies(idxhandler::DefaultIndexHandler)

Return indices of reduced species.

See also: elidedspecies

FiniteStateProjection.elidedspecies — Method

elidedspecies(idxhandler::DefaultIndexHandler)

Return indices of elided species.

See also: reducedspecies