Internal API

Index Handlers

Index Handler Interface

User-defined index handlers should inherit from AbstractIndexHandler and implement the following methods:

getsubstitutions

build_rhs_header

singleindices

pairedindices

FiniteStateProjection.singleindices — Method

singleindices(idxhandler::AbstractIndexHandler, arr)

Returns all indices I in arr. Defaults to CartesianIndices, but can be overloaded for arbitrary index handlers.

FiniteStateProjection.pairedindices — Function

pairedindices(idxhandler::AbstractIndexHandler, arr, shift::CartesianIndex)

Returns all pairs of indices (I .- shift, I) in arr.

FiniteStateProjection.getsubstitutions — Function

getsubstitutions(idxhandler::AbstractIndexHandler, sys::FSPSystem; state_sym::Symbol)::Dict

Construct the map speciesname => expr that gives the species abundances in terms of the state variable state_sym. See NaiveIndexHandler for the default implementation.

See also: build_ratefuncs, build_rhs

FiniteStateProjection.build_rhs_header — Method

build_rhs_header(idxhandler::AbstractIndexHandler, sys::FSPSystem)::Expr

Return initialisation code for the RHS function, unpacking the parameters p supplied by DifferentialEquations. The default implementation just unpacks parameters from p.

See also: unpackparams, build_rhs

Built-in implementations

FiniteStateProjection.elidedspecies — Method

elidedspecies(cons_laws::AbstractMatrix{Int})::Vector

Returns a list of species $[ s_1, ... ]$ which can be removed from the reaction system description using the provided matrix of conservation laws.

FiniteStateProjection.elisions — Function

elisions(idxhandler::DefaultIndexHandler, sys::FSPSystem)

Replaces the symbols $A(t)$, $B(t)$, ... of elided species by $N_1(t) - X(t) - Y(t)$, $N_2(t) - U(t) - V(t)$, ..., where $N_i(t)$ are the conserved quantities of the system.

See also: getsubstitutions

FiniteStateProjection.getsubstitutions — Method

getsubstitutions(idxhandler::NaiveIndexHandler, sys::FSPSystem; state_sym::Symbol)::Dict

Defines the abundance of species $S_i$ to be state_sym[i] - offset.

FiniteStateProjection.getsubstitutions — Method

getsubstitutions(idxhandler::DefaultIndexHandler, sys::FSPSystem; state_sym::Symbol)::Dict

Similar to its NaiveIndexHandler variant, but computes the abundances of elided species from the conserved quantities and the reduced species.

FiniteStateProjection.build_rhs_header — Method

build_rhs_header(idxhandler::DefaultIndexHandler, sys::FSPSystem)::Expr

Assumes p is of the form (params, cons::AbstractVector{Int}) where params are the system parameters and cons the conserved quantities.

FiniteStateProjection.pairedindices — Method

pairedindices(idxhandler::DefaultIndexHandler, arr::AbstractArray, shift::CartesianIndex)

Similar to its NaiveIndexHandler variant, but converts the indices into indices into the reduced state space array.

Function Building

FiniteStateProjection.build_rhs — Function

build_rhs(idxhandler::AbstractIndexHandler, sys::FSPSystem;
          combinatoric_ratelaw::Bool)

Builds the function f(du,u,p,t) that defines the right-hand side of the CME, for use in the ODE solver. If expression is true, returns an expression, else compiles the function.

FiniteStateProjection.unpackparams — Function

unpackparams(sys::FSPSystem, psym::Symbol)

Returns code unpacking the parameters of the system from the symbol psym in the form (p1, p2, ...) = psym. This should be called in all overloads of build_rhs_header. It is assumed that the variable psym is an AbstractVector{Float64}.

See also: build_rhs_header, build_rhs

FiniteStateProjection.build_ratefuncs — Function

build_ratefuncs(idxhandler::AbstractIndexHandler, sys::FSPSystem; 
                state_sym::Symbol, combinatoric_ratelaw::Bool)::Vector

Return the rate functions converted to Julia expressions in the state variable state_sym. Abundances of the species are computed using getsubstitutions.

See also: getsubstitutions, build_rhs

FiniteStateProjection.build_rhs_firstpass — Function

build_rhs_firstpass(sys::FSPSystem, rfs)::Expr

Return code for the first pass of the RHS function. Goes through all reactions and computes the negative part of the CME (probability flowing out of states). This is a simple array traversal and can be done in one go for all reactions.

See also: build_rhs

FiniteStateProjection.build_rhs_secondpass — Function

build_rhs_secondpass(sys::FSPSystem, rfs)::Expr

Return code for the second pass of the RHS function. Goes through all reactions and computes the positive part of the CME (probability flowing into states). This requires accessing du and u at different locations depending on the net stoichiometries. In order to reduce random memory access reactions are processed one by one.

See also: build_rhs