Main API

Reaction Systems

The following describes the FSPSystem struct, a thin wrapper around ModelingToolkit's ReactionSystem for use with this package. The main functionality of this type is determining conserved quantities for a reaction system.

FiniteStateProjection.FSPSystem — Type

struct FSPSystem
    rs::ReactionSystem
    cons_laws::Matrix{Int}
end

Thin wrapper around ModelingToolkit.ReactionSystem for use with this package. Automatically computes a matrix of conservation laws (see conservationlaws).

Constructor: FSPSystem(rs::ReactionSystem)

FiniteStateProjection.conservationlaws — Method

conservationlaws(sys::FSPSystem)::Matrix{Int}

Returns matrix of conservation laws associated with the system. Each row of the matrix contains the stoichiometric coefficients of a different conserved quantity.

FiniteStateProjection.conservedquantities — Function

conservedquantities(state, sys::FSPSystem)

Compute conserved quantities for the system at the given state.

Creating ODE systems

The following methods are the main way to create a system of ODEs representing the time-dependent FSP. This package provides a flexible way to represent the FSP in memory via index handlers, see [Index Handlers] for more information.

Base.convert — Method

convert(::Type{ODEFunction}, idxhandler::AbstractIndexHandler, sys::FSPSystem;
        combinatoric_ratelaw::Bool=true)

Return an ODEFunction defining the right-hand side of the CME.

Creates an ODEFunction for use with DifferentialEquations. This is where most of the work in the package happens; for best performance it is suggested to build an ODEFunction once for a given reaction system and reuse it instead of directly converting a reaction system to an ODEProblem (which implicitly calls this function).

Base.convert — Method

convert(::Type{ODEProblem}, idxhandler::AbstractIndexHandler, sys::FSPSystem, u0, tmax, p)

Return an ODEProblem for use in DifferentialEquations. This function implicitly callsconvert(ODEFunction, indexhandler, sys). It is usually more efficient to create anODEFunctionfirst and then use that to createODEProblem`s.