Matrix Conversions

FiniteStateProjection.jl provides functionality for building the right-hand side of the CME as a (sparse) matrix. This provides another way to solve the CME in time via

...

A = convert(SparseMatrixCSC, sys, idxhandler, dims, p)

prob = ODEProblem((du,u,p,t) -> mul!(du, p, u), u0, tt, A)

...

Note that the matrix A has to be rebuilt for every truncation size and every set of parameters, a restriction not shared by the ODEFunction returned by FiniteStateProjection.jl.

Base.convert — Method

Base.convert(::Type{SparseMatrixCSC}, ::FSPSystem, ::AbstractIndexHandler, dims::NTuple, ps;
             combinatoric_ratelaws=true)

Convert the reaction system into a sparse matrix defining the right-hand side of the Chemical Master Equation. dims is a tuple denoting the dimensions of the FSP and ps is the tuple of parameters. The sparse matrix works on the flattened version of the state obtained using vec.

Base.vec — Function

vec(idxhandler::AbstractIndexHandler, arr)

Converts the right-hand side defining the solution of the CME into a one-dimensional vector to which a matrix can be applied.

See also: LinearIndices